# workspace/fairsplit/api-spec.yaml
openapi: 3.1.0
info:
  title: FairSplit API
  version: 1.0.0
  description: |
    Zero-dependency shared expense tracker API.

    ## Authentication
    All group-scoped endpoints require a valid member session token.
    The token is issued by `POST /api/v1/groups` (group creation) or
    `POST /api/v1/groups/{group_id}/members` (join via invite link).

    **Primary:** `httpOnly` cookie named `fairsplit_member_token`.
    **Fallback:** `Authorization: Bearer <token>` header (API clients).

    The token is a JWT signed with HS256 containing:
    ```json
    {
      "sub": "<member_id>",
      "group_id": "<group_id>",
      "is_admin": true,
      "iat": 1713200000,
      "exp": 1715792000
    }
    ```

    ## Response Envelopes
    **Singleton responses** (auth, single-resource GET/POST/PATCH/DELETE):
    The resource object is returned directly at the top level. No wrapper.

    **List responses** (collection GET endpoints):
    ```json
    {
      "data": [...],
      "total": 23,
      "page": 1,
      "per_page": 20
    }
    ```

    **Error responses** (all 4xx and 5xx):
    ```json
    {
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Human-readable description.",
        "field": "amount",
        "request_id": "req_abc123"
      }
    }
    ```

    ## Idempotency
    Mutation endpoints (`POST` for create, `DELETE` for soft-delete) accept an
    `X-Idempotency-Key` header containing a client-generated UUID. The server
...

  contact:
...
servers:
...
paths:
...
  /groups:
    post:
      operationId: createGroup
      summary: Create a new expense group
      description: |
        Creates a group, adds the creator as the first member with the admin
        role, and issues a signed member session token (returned in the response
        body AND set as an httpOnly cookie).

        Idempotent: provide `X-Idempotency-Key` to prevent double-submission.
        A duplicate key returns the original 201 response without creating a
        second group.

        **Rate limit:** 10 requests per IP per hour.
      tags: [groups]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
            example:
              name: "Nashville Trip"
              description: "June 2026 friend group"
              admin_display_name: "Maya"
              default_currency: "USD"
      responses:
        "201":
          description: Group created successfully
          headers:
            Set-Cookie:
              description: |
                httpOnly session cookie. Format:
                fairsplit_member_token=<jwt>; Path=/; HttpOnly; Secure; SameSite=Lax; Max-Age=2592000
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateGroupResponse"
              example:
                group:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  name: "Nashville Trip"
                  description: "June 2026 friend group"
                  invite_token: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                  status: active
                  default_currency: USD
                  member_count: 1
                  expense_count: 0
                  created_at: "2026-04-15T14:00:00Z"
                  updated_at: "2026-04-15T14:00:00Z"
                member:
                  id: "a3bb1891-1f4e-4e7a-8c4f-6d4a1e2b3c4d"
                  display_name: "Maya"
                  is_admin: true
                  joined_at: "2026-04-15T14:00:00Z"
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimitExceeded"
        "500":
          $ref: "#/components/responses/InternalError"
  /groups/{group_id}:
    get:
      operationId: getGroup
      summary: Get group details
      description: |
        Returns group metadata including member count, expense count, and the
        invite token (for sharing the invite link).
        Requires a valid member session token for the specified group.
      tags: [groups]
      security:
        - MemberCookie: []
        - BearerToken: []
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                name: "Nashville Trip"
                description: "June 2026 friend group"
                invite_token: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
                status: active
                default_currency: USD
                member_count: 6
                expense_count: 23
                created_at: "2026-04-15T14:00:00Z"
                updated_at: "2026-04-15T18:30:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: updateGroup
      summary: Update group name, description, or status
      description: |
        Admin-only. Updates the group name, description, or status (active/archived).
        All fields are optional — only supplied fields are updated.
        Closing a group with outstanding balances returns a 409 with
        `GROUP_HAS_OUTSTANDING_BALANCES`. The admin may override with
        `force: true`.
      tags: [groups]
      security:
        - MemberCookie: []
        - BearerToken: []
      parameters:
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGroupRequest"
            example:
              name: "Nashville Trip 2026"
              status: archived
              force: false
      responses:
        "200":
          description: Group updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Group has outstanding balances (force=false)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: GROUP_HAS_OUTSTANDING_BALANCES
                  message: "Some members still have outstanding balances. Pass force=true to archive anyway."
                  request_id: req_abc123
        "422":
          $ref: "#/components/responses/ValidationError"
  /groups/{group_id}/invite:
    post:
      operationId: regenerateInviteToken
      summary: Regenerate the invite token (admin only)
      description: |
        Admin-only. Generates a new invite token for the group, invalidating
        the previous invite link. Useful if the invite link has been shared
        publicly and the admin wants to restrict new joins.
        Not needed in typical use — the default invite token is permanent.
      tags: [groups]
      security:
        - MemberCookie: []
        - BearerToken: []
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: New invite token generated
          content:
            application/json:
              schema:
                type: object
                required: [invite_token]
                properties:
                  invite_token:
                    type: string
                    format: uuid
                    description: New invite token. Old token is now invalid.
              example:
                invite_token: "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ============================================================
  # MEMBERS
  # ============================================================
...
components:
  securitySchemes:
    MemberCookie:
      type: apiKey
      in: cookie
      name: fairsplit_member_token
      description: |
        httpOnly session cookie issued on group creation or group join.
        Set by the server. Cannot be accessed by JavaScript.
        Expiry: 30 days, rolling (reset on each valid request).
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Fallback for API clients that cannot use cookies.
        Same JWT as the cookie value. Pass in Authorization: Bearer <token> header.
  parameters:
    GroupId:
      name: group_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: The group's UUID
...
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: |
        Client-generated UUID for idempotency protection. Generate one UUID
        when the form opens and reuse it on all retries. The server returns
        the original response for duplicate keys (within 24 hours).
  responses:
    Unauthorized:
      description: No session or session expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            no_session:
              value:
                error:
                  code: NO_SESSION
                  message: "No session token found. Please join the group first."
                  request_id: req_abc123
            session_expired:
              value:
                error:
                  code: SESSION_EXPIRED
                  message: "Your session has expired. Please re-enter your name to continue."
                  request_id: req_abc123
    Forbidden:
      description: Valid session but insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            permission_denied:
              value:
                error:
                  code: PERMISSION_DENIED
                  message: "You can only edit expenses you logged, or you must be the group admin."
                  request_id: req_abc123
            wrong_group:
              value:
                error:
                  code: MEMBER_WRONG_GROUP
                  message: "Your session token is for a different group."
                  request_id: req_abc123
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: GROUP_NOT_FOUND
              message: "Group not found. The link may be invalid."
              request_id: req_abc123
    ValidationError:
      description: Request body failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: "Amount must be greater than zero."
              field: amount_cents
              request_id: req_abc123
    RateLimitExceeded:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RATE_LIMIT_EXCEEDED
              message: "Too many requests. Please wait 42 seconds."
              request_id: req_abc123
    InternalError:
      description: Unhandled server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: "Something went wrong. Please try again."
              request_id: req_abc123
...
  schemas:
...
    Group:
      type: object
      required:
        - id
        - name
        - invite_token
        - status
        - default_currency
        - member_count
        - expense_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Group UUID (not the invite token — not guessable by design)
        name:
          type: string
          minLength: 1
          maxLength: 80
        description:
          type: string
          maxLength: 200
          nullable: true
        invite_token:
          type: string
          format: uuid
          description: |
            The token used in the invite link: {base_url}/join/{invite_token}.
            Different from the group ID — can be regenerated by admin.
        status:
          type: string
          enum: [active, archived]
        default_currency:
          type: string
          pattern: "^[A-Z]{3}$"
          example: USD
        member_count:
          type: integer
          minimum: 1
        expense_count:
          type: integer
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Member:
      type: object
      required: [id, display_name, is_admin, joined_at]
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
          minLength: 1
          maxLength: 40
        is_admin:
          type: boolean
        joined_at:
          type: string
          format: date-time
...
    CreateGroupRequest:
      type: object
      required: [name, admin_display_name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
          description: Group name. Whitespace-only values are rejected.
        description:
          type: string
          maxLength: 200
          nullable: true
        admin_display_name:
          type: string
          minLength: 1
          maxLength: 40
          description: Display name for the group creator (becomes the admin member)
        default_currency:
          type: string
          pattern: "^[A-Z]{3}$"
          default: "USD"
    CreateGroupResponse:
      type: object
      required: [group, member, token]
      properties:
        group:
          $ref: "#/components/schemas/Group"
        member:
          $ref: "#/components/schemas/Member"
        token:
          type: string
          description: |
            Signed JWT. Also set as an httpOnly cookie.
            Store in localStorage as a read cache (not the authoritative store).
            The cookie is the authoritative session store.
    UpdateGroupRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
          nullable: true
        description:
          type: string
          maxLength: 200
          nullable: true
        status:
          type: string
          enum: [active, archived]
          nullable: true
        force:
          type: boolean
          default: false
          description: |
            If true, archive the group even if some members have outstanding
            balances. Default: false (returns 409 if balances are nonzero).
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - GROUP_NOT_FOUND
                - MEMBER_NOT_FOUND
                - EXPENSE_NOT_FOUND
                - SETTLEMENT_NOT_FOUND
                - INVALID_INVITE_TOKEN
                - GROUP_ARCHIVED
                - GROUP_HAS_OUTSTANDING_BALANCES
                - PERMISSION_DENIED
                - NO_SESSION
                - SESSION_EXPIRED
                - SESSION_INVALID
                - MEMBER_WRONG_GROUP
                - IDEMPOTENCY_CONFLICT
                - SPLIT_SUM_MISMATCH
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
            message:
              type: string
              description: Human-readable error description
            field:
              type: string
              nullable: true
              description: The specific field that failed validation (for VALIDATION_ERROR only)
            request_id:
              type: string
              description: Unique request identifier for log correlation